+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # YTech Changelog
 
-## []
+## [0.10.1]
 
 - Fixed Strainer doesn't drop self